+++ src/library/DIPS.Mobile.UI/Components/Alerting/Dialog/dotnet/DialogService.cs
@@ -3,6 +3,15 @@ namespace DIPS.Mobile.UI.Components.Alerting.Dialog;
 
 public static partial class DialogService
 {
+    public static partial Task<InputDialogAction> ShowInputDialog(Action<IInputDialogConfigurator> configurator)
+    {
+        return Task.FromResult(new InputDialogAction
+        {
+            DialogAction = DialogAction.TappedAction,
+            DialogInputs = {  }
+        });
+    }
+    
     public static partial Task<DialogAction> ShowMessage(string title, string message, string actionTitle)
     {
         return Task.FromResult(DialogAction.TappedAction);
@@ -20,6 +29,11 @@ public static partial class DialogService
         return Task.FromResult(DialogAction.TappedAction);
     }
 
+    public static partial Task<DialogAction> ShowMessage(Action<IDialogConfigurator> configurator)
+    {
+        return Task.FromResult(DialogAction.TappedAction);
+    }
+    
     public static partial Task Remove()
     {
         return Task.FromResult(DialogAction.TappedAction);