+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## [42.2.3] 
+- [CollectionView][iOS] Fixed poor performance.
+
+## [42.2.2] 
+- [Chip][Android] Fixed so that default CornerRadius will be set.
+
+## [42.2.1] 
+- [SearchPage][ItemPicker] Fixed so that items will get correct default margin and corner radius
+
+## [42.2.0] 
+- Added Input dialog
+- Refactored DialogService
+
+## [42.1.1]
+- [CollectionView][Android] Fix bug where `AutoCornerRadius` would apply to header and footer.
+
 ## [42.1.0]
 - Added new global property to `Layout`: AutoCornerRadius. Will set corner radius on elements automatically, in lists, the first and last item in the list will receive corner radius.
 - [CollectionView] Added padding property.